fix some styles and little refactoring
