+++ src/components/RecipePageHero/RecipeIngredientsItem/RecipeIngredientsItem.module.css
@@ -81,6 +81,8 @@
     .measure {
         font-size: 16px;
         line-height: calc(20 / 16);
+        max-width: 112px;
+        text-overflow: ellipsis;
     }
   }
