+++ src/components/RecipePageHero/RecipeIngredientsList/RecipeIngredientsList.module.css
@@ -18,7 +18,7 @@
   .ingredientsList {
     gap: 22px;
   }
-  
+
   .title {
     margin-bottom: 24px;
   }
