+++ src/components/RecipePageHero/RecipePageHero.module.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   margin: 0 auto;
+
 }
   
 .textGlass {
@@ -10,7 +11,6 @@
   font-size: 12px;
   font-weight: 400;
   line-height: calc(14 / 12);
-
   color: rgba(243, 243, 243, 0.5);
 }
 
@@ -25,11 +25,15 @@
 }
 
 .imgDrink {
+  display:block;
   min-width: 335px;
-  height: 400px;
+  /* height: 400px; */
+  height: auto;
   margin-bottom: 40px;
   border-radius: 8px;
   object-fit: cover;
+  box-shadow: 0px 0px 40px #ffffff18;
+  object-position: top center;
 }
 
 @media screen and (min-width: 768px) {    
@@ -47,13 +51,18 @@
   
   .imgDrink {
     min-width: 704px;
+    max-height: 400px;
     margin-bottom: 80px;
+
   }
 }
+
+
 
 @media screen and (min-width: 1440px) {
   .sectionContainer {
     flex-direction: row;
+
   }
   
   .heroContainer {
@@ -63,5 +72,6 @@
   .imgDrink {
     min-width: 400px;
     margin-bottom: 100px;
+
   }
 }
