+++ src/components/RecipePageHero/RecipePreparation/RecipePreparation.module.css
@@ -18,9 +18,10 @@
   min-width: 335px;
   height: 430px;
   border-radius: 8px;
-  
+
   transform: scaleX(-1);
   object-fit: cover;
+  object-position: top center;
 }
 
 .description {
@@ -92,6 +93,7 @@
   .recipeContainer {
     flex-direction: row;
     gap: 60px;
+
   }
 
   .title {
