+++ src/pages/RecipePage/RecipePage.jsx
@@ -7,9 +7,9 @@ import useAuth from 'hooks/useAuth';
 import { nanoid } from '@reduxjs/toolkit';
 
 import RecipePageHero from 'components/RecipePageHero/RecipePageHero';
-import RecipeIngredientsList from 'unused/RecipeIngredientsList/RecipeIngredientsList';
-import RecipeIngredientsItem from 'unused/RecipeIngredientsItem/RecipeIngredientsItem';
-import RecipePreparation from 'components/RecipePreparation/RecipePreparation';
+import RecipeIngredientsList from 'components/RecipePageHero/RecipeIngredientsList/RecipeIngredientsList';
+import RecipeIngredientsItem from 'components/RecipePageHero/RecipeIngredientsItem/RecipeIngredientsItem';
+import RecipePreparation from 'components/RecipePageHero/RecipePreparation/RecipePreparation';
 import { getFavPage } from 'redux/FavoriteCocktails/FavoritesSelectors';
 import { addFavorites, deleteFavorites, fetchFavorites } from 'redux/FavoriteCocktails/FavoritesOperation';
